fixing launcher package
