+++ packages/launcher/src/index.ts
@@ -1,15 +1,2 @@
-export function getTreasureLauncherAuthToken(): string | undefined {
-  const windowAuthToken = window.location.search.match(
-    /--tdk-auth-token=(\w+)/,
-  )?.[1];
-  const processBackendWallet = process.argv
-    .find((arg) => arg.startsWith("--tdk-auth-token="))
-    ?.split("=")[1];
-  return windowAuthToken || processBackendWallet;
-}
-
-export function isUsingTreasureLauncher(): boolean {
-  return !!getTreasureLauncherAuthToken();
-}
-
+export { getTreasureLauncherAuthToken, isUsingTreasureLauncher } from "./utils";
 export { startUserSessionViaLauncher } from "./session";
