+++ packages/launcher/src/session.ts
@@ -1,5 +1,6 @@
 import type { SessionOptions } from "@treasure-dev/tdk-core";
 import axios from "axios";
+import { isUsingTreasureLauncher } from "./utils";
 
 export function startUserSessionViaLauncher({
   backendWallet,
@@ -8,6 +9,13 @@ export function startUserSessionViaLauncher({
   sessionDurationSec,
   sessionMinDurationLeftSec,
 }: SessionOptions): Promise<void> {
+  if (!isUsingTreasureLauncher()) {
+    return Promise.reject(
+      new Error(
+        "startUserSessionViaLauncher can only be used with Treasure Launcher",
+      ),
+    );
+  }
   return axios.post("http://localhost:16001/tdk-start-session", {
     backendWallet,
     approvedTargets,
